Fixed more bad code/improved leaping across empty spaces code
Fixed bad code that would cause Nemeses to not leap across empty spaces to get to the player. They can now jump longer distances, with a little more tolerance on the y axis. Be careful if there is an almost flat gap between you and the Nemesis.

Gave *slightly* more tolerance for Nemeses building up to the player.

Fixed a bug which caused Nemeses with Overwhelming Will to not damage nearby entities.