Better AI (opening doors, taking abilities into context), performance, and abilities update.dsc
Nemeses which aren't Drowned mobs will usually try to avoid water when pathfinding to you or an enemy

All Nemeses should now be able to open doors

Nemeses with the ability NoFallDamage will now adjust their pathfinding according to the ability. For example, if the Nemesis is on top of a cliff and you are at the bottom of the cliff, the Nemesis will not be afraid to fall down multiple blocks to get to you since they are immune to fall damage.

Added ImmunePoison ability for Nemeses. If a Nemesis learns this ability, they are immune to poison attacks.

Added ImmuneWither ability for Nemeses. If a Nemesis learns this ability, they are immune to wither attacks.

For performance reasons, Nemeses damaged by fire will have a lower pathfinding range to extinguish themselves.